py-client/pyproject.toml
[tool.poetry]
name = "py-client"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.7"
pydantic = {extras = ["email"], version = "^1.8.2"}
requests = "^2.26.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
commitizen = "^2.18.1"
pdoc3 = "^0.10.0"
rich = "^10.11.0"
python-dotenv = "^0.19.0"
fire = "^0.4.0"
autopep8 = "^1.5.7"

[tool.poetry.scripts]
generate = 'scripts.generate:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
